[WIP][InferAtomsPass] Instruction scheduling
Step 1 of optimizing atomic regions for (smaller) size.

In essence, it's now necessary to have a complete picture of which
instructions are tainted (whereas before we really only needed to know
the boundaries of a region).

Test plan: `make eg3` for an example where the freshness atomic region
size is reduced thanks to the optimization.